time series db adaptor/requirements.txt
CherryPy==18.10.0
mysql_connector_repackaged==0.3.1
paho_mqtt==1.6.1
Requests==2.32.3
